+++ visualization/app/codeCharta/views/metricsView/explorer/metricsExplorerRules.ts
@@ -2,7 +2,7 @@ import { Injectable, inject } from "@angular/core"
 import { Store } from "@ngrx/store"
 import { ExplorerRules, RuleWithCount } from "../../../features/sidebarExplorer/facade"
 import { BlacklistType, CcState } from "../../../model/codeCharta.model"
-import { removeBlacklistItem } from "../../../stores/sharedView/sharedView.write.facade"
+import { removeBlacklistItem, removeMetricRule } from "../../../stores/sharedView/sharedView.write.facade"
 import { dispatchAfterPaint } from "../../../util/dispatchAfterPaint"
 import { blacklistSearchPattern } from "../effects/blacklistSearchPattern/blacklistSearchPattern.effect"
 import { excludeRulesWithCountSelector, flattenRulesWithCountSelector } from "./explorerRules.selectors"
@@ -18,6 +18,10 @@ export class MetricsExplorerRules implements ExplorerRules {
     readonly isExcludePatternDisabled$ = this.store.select(isExcludePatternDisabledSelector)
 
     removeRule(rule: RuleWithCount) {
+        if (rule.kind === "METRIC") {
+            dispatchAfterPaint(this.store, removeMetricRule({ id: rule.metricRule.id }))
+            return
+        }
         dispatchAfterPaint(this.store, removeBlacklistItem({ item: rule.item }))
     }